Changed the arrival time data type handling. Now the max amplitude positions are int, cause they are used also as array indices. The processors which calculate a arrivalTime (RisingEdge and TimeOverThreshold) put a double array in the data item. All following processor now expect a double array. Also adapted the unit tests

+++ src/main/java/fact/extraction/MaxAmplitudeInSearchWindow.java
@@ -20,7 +20,7 @@ public class MaxAmplitudeInSearchWindow implements Processor {
 	private int searchWindowRight;
 	
 	private double[] amplitudes = new double[Constants.NUMBEROFPIXEL];
-	private double[] positions = new double[Constants.NUMBEROFPIXEL];
+	private int[] positions = new int[Constants.NUMBEROFPIXEL];
 	
 	int roi;
